resume_api/requirements.txt
pyOpenSSL
Flask
numpy
pandas
spacy==2.3.0
pdfminer.six
docx2txt
tika
nltk
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz#egg=en_core_web_sm